+++ src/app.d.ts
@@ -1,12 +1,14 @@
 // See https://kit.svelte.dev/docs/types#app
 // for information about these interfaces
 declare global {
-	namespace App {
-		// interface Error {}
-		// interface Locals {}
-		// interface PageData {}
-		// interface Platform {}
-	}
+  namespace App {
+    // interface Error {}
+    // interface Locals {}
+    interface PageData {
+      flash?: { type: 'success' | 'error'; message: string };
+    }
+    // interface Platform {}
+  }
 }
 
 export {};