+++ CHANGELOG.md
@@ -5,6 +5,16 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2023-02-23
+
+### Added
+
+- Support for [sveltekit-flash-message](https://www.npmjs.com/package/sveltekit-flash-message)
+
+### Changed
+
+- `options.defaultValidator` now has default value `clear`.
+
 ## [0.4.0] - 2023-02-22
 
 ### Changed